feat: list is now fully functional

+++ src/components/ActivityList/index.tsx
@@ -1,27 +1,67 @@
-import { useContext, useEffect, useState } from 'react';
+import { useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { getActivities } from '../../services/getActivities';
 import { Activity as ActivityType } from '../../types/Activity';
-import Activity from '../Activity';
 import * as S from './styles';
 
 import { ReactComponent as DeleteIcon } from '../../assets/icon-cross.svg';
 import { ModesContext } from '../../context/ModesContext';
 import { ActivitiesContext } from '../../context/ActivityContext';
 import { deleteActivity } from '../../services/deleteActivity';
+import Activity from '../Activity';
 
 export function ActivityList() {
   const {mode} = useContext(ModesContext);
-  const {activities, setActivities} = useContext(ActivitiesContext);
+  const {activities, setActivities, active, setActive} = useContext(ActivitiesContext);
+  const [displayActivity, setDisplayActivity] = useState<ActivityType[] | null>(null);
+  const [counter, setCounter] = useState<number>(0);
 
-  const handleDelete = async (id: string) => {
+
+  const handleActive = (current: number) => {
+    setActive(current);
+
+    if (activities) {
+      if (current === 0) {
+        setDisplayActivity(activities);
+      } else if (current === 1) {
+        setDisplayActivity(activities.filter(data => data.status === 0));
+      } else {
+        setDisplayActivity(activities.filter(data => data.status === 1));
+      }
+    }
+  };
+
+  const handleDelete = async (id: string, canBeDeleted: boolean) => {
+    if (!canBeDeleted) {
+      alert('Design messages cannot be deleted.');
+      return;
+    }
+
     await deleteActivity(id).then(() => {
       getActivities().then(response => setActivities(response));
     });
   };
 
+  const countLeftItems = () => {
+    let count = 0;
+    if (activities) {
+      activities.map(data => {
+        if (data.status === 0) {
+          count++;
+        }
+      });
+      setCounter(count);
+    }
+  };
+
+  useEffect(() => {
+    setDisplayActivity(activities);
+    countLeftItems();
+  }, [activities]);
+
   useEffect(() => {
     getActivities().then((response) => {
       setActivities(response);
+      setDisplayActivity(response);
     });
   }, []);
 
@@ -30,12 +70,12 @@ export function ActivityList() {
       <S.ListContainer>
         <S.List>
           {
-            activities ? activities.map((activity) => (
+            displayActivity ? displayActivity.map((activity) => (
               <Activity key={activity._id} status={activity.status} id={activity._id}>
                 <S.Text>{activity.message}</S.Text>
                 <DeleteIcon
                   stroke={mode ? 'hsl(233, 14%, 35%)' : 'hsl(235, 19%, 35%)'}
-                  onClick={() => handleDelete(activity._id)}/>
+                  onClick={() => handleDelete(activity._id, activity.isDeletable)}/>
               </Activity>
             )) : (
               <Activity>
@@ -47,19 +87,28 @@ export function ActivityList() {
 
         <S.Footer>
           <S.FooterContent>
-            <span>{activities?.length} items left</span> {/* temporary */}
+            <span>{counter} items left</span>
           </S.FooterContent>
 
           <S.FooterContent>
-            <S.Button active={true} weight='700'>
+            <S.Button
+              active={active === 0 ? true : false}
+              weight='700'
+              onClick={() => handleActive(0)}>
               All
             </S.Button>
 
-            <S.Button weight='700'>
+            <S.Button
+              active={active === 1 ? true : false}
+              weight='700'
+              onClick={() => handleActive(1)}>
               Active
             </S.Button>
 
-            <S.Button weight='700'>
+            <S.Button
+              active={active === 2 ? true : false}
+              weight='700'
+              onClick={() => handleActive(2)}>
               Completed
             </S.Button>
           </S.FooterContent>
